Change entry point to state sorter
